+++ wishFit/src/main/java/wishFit/servlet/member/MemberProfileServlet.java
@@ -16,18 +16,19 @@ import wishFit.beans.member.MemberProfileDao;
 import wishFit.beans.member.MemberProfileDto;
 
 
-@WebServlet(urlPatterns="/profile.kh")
+@WebServlet(urlPatterns="/page/fitgroup/profile.kh")
 public class MemberProfileServlet extends HttpServlet{
 	@Override
 	protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
 		
 		try {
 			int mpNo  = Integer.parseInt(req.getParameter("mpNo"));
-
+			System.out.println("1");
 			MemberProfileDao memberProfileDao = new MemberProfileDao();
 			MemberProfileDto memberProfileDto = memberProfileDao.get(mpNo);
+			System.out.println("2");
 
-			File dir = new File("D:\\upload\\member");
+			File dir = new File("D:\\upload\\wishfit");
 			File target = new File(dir, memberProfileDto.getMpSave());
 			InputStream in = new FileInputStream(target);
 			byte[] buffer = new byte[8192];
